warehouse-api: small code refactor
