+++ DOS/warehouse-api/service/src/main/java/com/immortalcrab/warehouse/verticles/Principal.java
@@ -1,6 +1,6 @@
 package com.immortalcrab.warehouse.verticles;
 
-import com.immortalcrab.warehouse.endpoints.Transfers;
+import com.immortalcrab.warehouse.applications.Transfers;
 import io.vertx.config.ConfigRetriever;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.DeploymentOptions;
