+++ DOS/warehouse-api/service/src/main/java/com/immortalcrab/warehouse/verticles/SyncDbBridge.java
@@ -1,6 +1,6 @@
 package com.immortalcrab.warehouse.verticles;
 
-import com.immortalcrab.warehouse.endpoints.Transfers;
+import com.immortalcrab.warehouse.applications.Transfers;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.eventbus.EventBus;
 import io.vertx.core.json.JsonObject;
